Update for namespaced services in hwservicemanager.
IServiceManager now takes interfaceChain for registration and a
fully-qualified name for service retrieval. This allows for services to
register as their parent classes as well. So, for instance, a vendor hal
can serve both vendor code and the android framework which expects a
non-customized hal.

Technically, we should be able to call 'interfaceChain' on the interface
that is passed into the service manager, however at the time that 'get'
is called, the registering hal won't be serving itself over binder.

Test: end to end
Bug: 31861268
Change-Id: I9904b3ab959a47a5556d5773ffe6fb535983cf00