Return<T>: fail hard on unchecked error.
If a transaction error occurs during a hidl call, and it isn't handled,
we want to fail hard, fast, and early. This adds a check in the Return
destructors which does this.

As part of this, I also combined the common parts of Return<*> and its
specialization Return<void> into a parent type, details::return_status.

Bug: 31348667
Test: hidl_test
Change-Id: Iba6d43ffb2e1dc68ba2c291f6f0c5ef1257e1da9