memory service: Remove memory@1.0-impl from req'd.
Requiring a package requires the package, but only the parts of it that
could be used by a specific package. So, the 32 bit version was not
being built, and explicitly requiring "android.hidl.memory@1.0-impl-32"
would not make it be built either. Instead android.hidl.memory@1.0-impl
has to be added to PRODUCT_PACKAGES on the device. This is done in
embedded.mk.

Additionally, specified compile_multilib: "both" to be explicit that
both should be built (although it appears as if this is the default).

Fix: 34134179
Test: clean build has both 32 and 64 bit versions of the -impl on the
device.

Change-Id: Ie88ef08eefd05f78962db44c0d46cfa3a4d8d9bd