Add waitForHwService.
This is a helper method which will wait for a service that is supposed
to be binderized. This is an implementation detail of getService and is
only called when the VINTF indicates a service is supposed to be
binderized.

Note this is essentially a no-op because the only services which are
marked as hwbinder are Java services (which have a different getService
code path). I did test this code path, however, with nfc as specified in
the first test line.

Bug: 31458381
Test: marked nfc as hwbinder in the VINTF, and its start was
successfully delayed.
Test: ENABLE_TREBLE=false, manual tests of fingerprint/wifi/nfc/lights.
No associated log messages.
Test: ENABLE_TREBLE=true, same tests and results as above.

Change-Id: I373761df624ebf60661726bb81d979a8c7289b1d