return_status::checkStatus => assertOk
checkStatus doesn't really set mCheckStatus to true, and hence is
a misnomer. Change it to assertOk (correspond to isOk()).

And hence the move assignment operator should not call checkStatus
(because it is valid to move into a Return<T> object that !isOk()
but has mCheckedStatus to true), because caller is fully aware of
the error that will be overwritten.

Test: libhidl_test
Change-Id: I86a44967b68619d1467d2cc9caaa39124b156121