Mark hidl libs as VNDK or VNDK-SP in Android.bp
As a VNDK/VNDK-SP module, Android.bp must have 'vndk' tag as well
as 'vendor_available: true'.

For a VNDK module, the 'vndk' tag has 'enabled: true'.
It will be installed system/lib(64)/vndk as a vendor variant.

For a VNDK-SP module, the 'vndk' tag has
'support_system_process: true' as well as 'enabled: true'.
It will be installed system/lib(64)/vndk-sp as a vendor variant.

Note: libhidlmemory will be changed to vndk-sp.

Bug: 63866913
Test: build and boot with BOARD_VNDK_VERSION=current
Change-Id: I9e3c9603f1d218c17ccb8e6aea7afad6f9d28b92