Mark android.hidl.memory@1.0-impl as vndk-sp
This framework HAL is marked as vndk-sp since it is depended by
libhidlmemory which is a vndk-sp library. Unlike other HALs from
vendors, this HAL can't be placed under /vendor/lib/hw since it is
'framework' HAL that is owned by the platform owner. In order to update
the HAL via platform-only update, the lib is marked as vndk-sp, but not
as a vendor-only lib.

Bug: 62930720
Test: BOARD_VNDK_VERSION=current m android.hidl.memory@1.0-impl
successful
Test: wahoo devices boots

Change-Id: I7fefb3d3f14ad833afe5cf0288b4d0c3fca1f9c8